msm8916-common: remove android.hardware.bluetooth.audio because it's not common to all msm8916 tablets
The SM-T350 uses a proprietary sound driver for the amplifier and only works with audio 2.0.
This has been the case since the beginning and not specific to Android 12.

So we need to put this in the specific device tree and not in msm8916-common.

If android.hardware.bluetooth.audio is present for SM-T350, audio over bluetoot doesn't work.